update MethodOverload
changed float to decimal

+++ MethodOverloading/MethodOverloading/mathFunction.cs
@@ -13,7 +13,7 @@ namespace MethodOverloading
         }
 
         //this method multiplies 5 to user's float/decimal input
-        public static int myOverload(float num)
+        public static int myOverload(decimal num)
         {
             int myNum = Convert.ToInt32(num);
             int myNumMult = myNum * 5;
